refactor: moves banner into cards
